requirements.txt
flask
flask-cors
pytesseract
pillow
PyMuPDF
